+++ crates/whisker-types/src/decorated_tree.rs
@@ -47,7 +47,7 @@ impl DecoratedTree {
             self.tree.root_node(),
             &self.source,
             &self.file,
-            &self.decorations,
+            self.decorations.lookup(),
         )
     }
 
@@ -82,7 +82,7 @@ mod tests {
     #[derive(Eq, PartialEq, Debug)]
     struct Tag(u32);
 
-    unsafe impl Decoration for Tag {
+    impl Decoration for Tag {
         const KEY: DecorationKey = DecorationKey::new(concat!(module_path!(), "::Tag"));
 
         type Ref<'a> = Option<&'a Self>;